include verhengt huso

+++ Software/Cable_Monitor_Code/Core/Src/measuring.c
@@ -61,6 +61,7 @@
 #include "stm32f4xx.h"
 
 #include "measuring.h"
+#include "calculation.h"
 
 /******************************************************************************
  * Defines
@@ -354,7 +355,7 @@ void MEAS_show_data(void) {
     char text[32];
 	// Calculate the frequency of channel 1
 	static float frequency = 0;
-    frequency = calculate_main_frequency(1, Samples, INPUT_COUNT, ADC_NUMS, ADC_FS);
+	frequency = calculate_main_frequency(1, Samples, INPUT_COUNT, ADC_NUMS, ADC_FS);
     //uint32_t frequency = frequency_calculation(0, Samples);
     // Calculate the peak value of channel 1
     uint32_t peak = 0;
@@ -365,7 +366,7 @@ void MEAS_show_data(void) {
     }
     peak -= 2048;
     BSP_LCD_SetTextColor(LCD_COLOR_CYAN);
-    snprintf(text, 40, "C1:%4d f:%3f HZ", (int)(peak), (frequency));
+    snprintf(text, 40, "C1:%4d f:%3d HZ", (int)(peak), (int)(frequency));
     // snprintf(text, 15, "C1 %4d", (int)(Samples[0]));
     BSP_LCD_DisplayStringAt(0, 140, (uint8_t*)text, LEFT_MODE);
 	// Calculate the frequency of channel 2
@@ -410,7 +411,7 @@ void MEAS_show_data(void) {
     }
     peak -= 2048;
     BSP_LCD_SetTextColor(LCD_COLOR_GREEN);
-    snprintf(text, 40, "C4:%4d f:%3d HZ", (int)(peak), (int)(frequency));
+    snprintf(text, 40, "C4:%4d f:%3d HZ", (int)(peak), (float)(frequency));
     // snprintf(text, 15, "C4 %4d", (int)(Samples[3]));
     BSP_LCD_DisplayStringAt(0, 230, (uint8_t*)text, LEFT_MODE);
     /* Draw the  values of input channel 1 as a curve */
